provider: Move Provider out of uppy-base
