+++ src/plugins/Instagram/index.js
@@ -1,9 +1,9 @@
 const html = require('yo-yo')
 const Plugin = require('../Plugin')
 
-const Provider = require('../../uppy-base/src/plugins/Provider')
+const Provider = require('../../Provider')
 
-const View = require('../../generic-provider-views/index')
+const View = require('../../generic-provider-views')
 
 module.exports = class Instagram extends Plugin {
   constructor (core, opts) {
